organized the body patterns into src/app/patterns/body/elements/ where they belong
