+++ src/app/patterns/body/elements/ComparisonSummary.svelte
